+++ src/SerialLogger.h
@@ -1,6 +1,7 @@
 #ifndef HUMIDISTAT_SERIALLOGGER_H
 #define HUMIDISTAT_SERIALLOGGER_H
 
+#include <stdint.h>
 #include "Humidistat.h"
 #include "ThermistorReader.h"
 
@@ -11,14 +12,14 @@ private:
 	ThermistorReader (&trs)[4];
 
 	const char *header = "Humidity Setpoint Temperature ControlValue T0 T1 T2 T3";
-	const unsigned long interval; //!< Logging interval (in millis)
+	const uint16_t interval; //!< Logging interval (in millis)
 
 	unsigned long lastTime = 0;   //!< Last time line was written (in millis)
 	bool ready = false;
 public:
 	/// Constructor.
 	/// \param humidistat Pointer to a Humidistat instance
-	explicit SerialLogger(Humidistat *humidistat, ThermistorReader (*trs)[4], unsigned long interval);
+	explicit SerialLogger(Humidistat *humidistat, ThermistorReader (*trs)[4], uint16_t interval);
 
 	/// Setup the serial interface
 	void begin();
